signature, ordering, static helper

+++ src/test/java/ch/ethz/idsc/tensor/alg/OrderingTest.java
@@ -19,7 +19,7 @@ import junit.framework.TestCase;
 public class OrderingTest extends TestCase {
   public void testVector() throws ClassNotFoundException, IOException {
     Tensor vector = Tensors.vector(4, 2, 3, 0, 1);
-    int[] array = Serialization.copy(Ordering.INCREASING).of(vector);
+    Integer[] array = Serialization.copy(Ordering.INCREASING).of(vector);
     Tensor ascending = Tensor.of( //
         IntStream.range(0, array.length).mapToObj(index -> vector.Get(array[index])));
     assertEquals(ascending, Sort.of(vector));
@@ -28,7 +28,7 @@ public class OrderingTest extends TestCase {
   public void testRandom() {
     Distribution d = BinomialDistribution.of(12, RationalScalar.of(1, 3));
     Tensor vector = RandomVariate.of(d, 1000);
-    int[] array = Ordering.INCREASING.of(vector);
+    Integer[] array = Ordering.INCREASING.of(vector);
     Tensor ascending = Tensor.of( //
         IntStream.range(0, array.length).mapToObj(index -> vector.Get(array[index])));
     assertEquals(ascending, Sort.of(vector));
@@ -37,7 +37,7 @@ public class OrderingTest extends TestCase {
   public void testNormal() {
     Distribution d = NormalDistribution.standard();
     Tensor vector = RandomVariate.of(d, 1000);
-    int[] array = Ordering.INCREASING.of(vector);
+    Integer[] array = Ordering.INCREASING.of(vector);
     Tensor ascending = Tensor.of( //
         IntStream.range(0, array.length).mapToObj(index -> vector.Get(array[index])));
     assertEquals(ascending, Sort.of(vector));
@@ -46,7 +46,7 @@ public class OrderingTest extends TestCase {
   public void testNormalDecreasing() {
     Distribution d = NormalDistribution.standard();
     Tensor vector = RandomVariate.of(d, 1000);
-    int[] array = Ordering.DECREASING.of(vector);
+    Integer[] array = Ordering.DECREASING.of(vector);
     Tensor decreasing = Tensor.of( //
         IntStream.range(0, array.length).mapToObj(index -> vector.Get(array[index])));
     assertEquals(Reverse.of(decreasing), Sort.of(vector));
@@ -65,9 +65,9 @@ public class OrderingTest extends TestCase {
 
   public void testSingleton() {
     Tensor tensor = Tensors.fromString("{{1, 2, 3}}");
-    int[] array = Ordering.DECREASING.of(tensor);
+    Integer[] array = Ordering.DECREASING.of(tensor);
     assertEquals(array.length, 1);
-    assertEquals(array[0], 0);
+    assertEquals(array[0].intValue(), 0);
   }
 
   public void testScalarFail() {
